Add files via upload
I'd like to propose the addition of an .everyN() traversal method to the jQuery UI project - you can think of it as a superset of the existing .odd() and .even() methods. I've tried to copy the jQuery description of these methods below to explain how .everyN() would work:

everyN(N, "skip")

Description: Reduce the set of matched elements to every Nth one in the set. Add the string "skip" as an optional parameter to exclude every Nth element from the total set of matched elements.